+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v0.9.2
 
 ### Changed